将公用代码抽取到不同层级的父类中
