+++ src/main/java/com/github/hcsp/inheritance/Animal.java
@@ -2,15 +2,9 @@ package com.github.hcsp.inheritance;
 
 public class Animal {
     private String name;
-    private int age;
 
     public Animal(String name) {
         this.name = name;
-    }
-
-    public Animal(String name, int age) {
-        this.name = name;
-        this.age = age;
     }
 
     public String getName() {
@@ -21,19 +15,8 @@ public class Animal {
         this.name = name;
     }
 
-    public int getAge() {
-        return age;
-    }
-
-    public void setAge(int age) {
-        this.age = age;
-    }
-
     public void sayMyName() {
         System.out.println("我的名字是" + name);
     }
 
-    public void sayMyAge() {
-        System.out.println("我的年龄是" + age);
-    }
 }
