+++ src/main/java/com/github/hcsp/inheritance/Boy.java
@@ -1,10 +1,9 @@
 package com.github.hcsp.inheritance;
 
-public class Boy extends Animal {
+public class Boy extends Person {
     public Boy(String name, int age) {
         super(name, age);
     }
-
 
     public void sayBoy() {
         System.out.println("我是一个男孩");
