+++ src/main/java/com/github/hcsp/inheritance/Girl.java
@@ -1,6 +1,6 @@
 package com.github.hcsp.inheritance;
 
-public class Girl extends Animal {
+public class Girl extends Person {
     public Girl(String name, int age) {
         super(name, age);
     }
